Add expected ADC values for ESP32 and AVR
Also move constants into own file to reduce clutter in header
